fix: split title detection helper below budget
